requirements.txt
numpy==1.19.5
scikit-learn==0.24.1
scipy==1.6.0
pandas==1.2.1
PyYAML==5.4.1
torch==1.7.1
